fix(assistants): honor current listener delta replacements
